+++ packages/generator-metaed-plugin/generators/app/templates/src/generator/DomainEntityNameGenerator.ts
@@ -1,4 +1,3 @@
-
 import { promisify } from 'util';
 import path from 'path';
 import fs from 'fs';