+++ packages/metaed-plugin-edfi-interchange-brief/index.ts
@@ -1,4 +1,4 @@
-// Reminder - this index.js is for on-the-fly transpile when there is no dist directory
+// Reminder - this index.ts is for on-the-fly transpile when there is no dist directory
 
-// All exports from the "real" index.js
+// All exports from the "real" index.ts
 export * from './src/index';