+++ CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1-dev.7](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.6...v1.4.1-dev.7) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
+## [1.4.1-dev.6](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.5...v1.4.1-dev.6) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
+## [1.4.1-dev.5](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.4...v1.4.1-dev.5) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
+## [1.4.1-dev.4](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.3...v1.4.1-dev.4) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
+## [1.4.1-dev.3](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.2...v1.4.1-dev.3) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
+## [1.4.1-dev.2](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.0...v1.4.1-dev.2) (2019-01-24)
+
+**Note:** Version bump only for package metaed-js
+
+
+
+
+
 ## [1.4.1-dev.1](https://github.com/Ed-Fi-Alliance/MetaEd-js/compare/v1.4.1-dev.0...v1.4.1-dev.1) (2019-01-22)
 
 **Note:** Version bump only for package metaed-js